facet-toml: ser: TomlSerializer: move traces to avoid debug-only args
It feels strange to have functions take arguments that are only used in
debug contexts, and have no potential for future use. This commit
removes the `type_name` argument from `push_key` and `pop_key`. This
argument had a misleading name to begin with.

Instead, the traces are moved to a position that has all of the required context;
there is no further need to pass around static strings for debugging
purposes, keeping business logic clean.