facet-core: Shape: Debug: omit field vtable
For normal formatting, omit the field completely, and finish the debug
formatting as non-exhaustive (append `..` to the end of the fields).

For alternate formatting, since it's on multiple lines, there is very
little price to at least show that the field exists. For this reason, it
is retained, but because these pointers are unlikely to be useful, the
value itself is elided and kept to a single line.